Removed faulty assertion in MultipleOutputsChannelElement<T> and forward buffer_policy to base class in SharedConnection<T> constructor
The logic of the assertions in MultipleOutputsChannelElement<T>::data_sample(sample) and MultipleOutputsChannelElement<T>::write(sample)
was wrong. All instances of MultipleOutputsChannelElement<T> can have more than one output, but only in case of a
PerOutputPort policy the shared buffer is installed before the actual channel input and ConnInputEndPoint<T>::write(sample)
should never be called in this case. This was the intention of the original assertion added in afe68a89ea8fb5d4c631411cf2af56573610573a,
but the same is not true for buffer_policy == Shared connection and for data_sample(sample), which is forwarded even accross buffer
boundaries.

As a consequence of debugging this bug the assertion was triggered during the initial SharedConnection<T>::data_sample(sample)
call, because the SharedConnection<T> constructor did not properly set the buffer_policy member variable of its base class
MultipleInputsMultipleOutputsChannelElement<T> to Shared, as expected.

Both changes have no influence on the actual behavior because the buffer policy only controls the channel selection of
the MultipleInputsChannelElement<T> base class, but all relevant methods are reimplemented in SharedConnection<T>.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>