base: fixed and simplified DataObjectLockFree::Set() implementation
The previous implementation was not preventing that multiple writers enter the critical section concurrently,
only that they try to write to the same sample. If they did, there was a race condition in updating the read
and write pointers afterwards.

The new solution is much simpler and drops out early if the Set() call detects that another writer is active
concurrently, so it behaves similar to a critical section protected by a MutexTryLock.

The assertions for assignments in the dataflow_performance_test has been improved to cover the case where
the right-hand side of an assignment is the target of another assignment, which would result in inconsistent
copies. This assertion was triggered before without the patch for DataObjectLockFree.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>