Always try to read from the same channel in MultipleInputsChannelElement
Except for buffer policy `PerInputPort` every input port has to select a channel from the
list of connections for each read operation. Before this commit, the channel selection
strategy depended on the type of buffer policy and each `MultipleInputsChannelElement`
or `MultipleOutputsChannelElement` only allowed to set the buffer policy once. Or, in
other words, a single port could not have multiple connections with different buffer
policies.

This commit drops the `buffer_policy` member variable and changes the
`MultipleInputsChannelElement` reader selection to apply the "start with the
same channel that had new data the last time" unconditionally, which before was only
selected for buffer policies `PerConnection` and `PerOutputPort`. So the only case where
this makes a difference is for input ports with multiple connections of type `Shared`.
The `SharedConnection<T>` instance itself, which is also a `MultipleInputsChannelElement`,
overrides the read method and is not influenced by the channel selection either.

The channel selection strategy that always reads from the last signaled/written channel
introduced in afe68a89ea8fb5d4c631411cf2af56573610573a was only effective in the case of
an input port with multiple `Shared` connections.

On the positive side some restrictions on mixed buffer policies for connections to the
same port have been relaxed and a significant amount of code could be removed from the
channel element implementation. The section in `channelelements_test.cpp` that tested
the "last signaled" strategy and that has been added in
afe68a89ea8fb5d4c631411cf2af56573610573a also has been removed.

The `MultipleInputsChannelElement::currentInput()` member function, which returns a
pointer to the last selected channel, is moved to the private section as it is only
used internally and not necessarily thread-safe in other contexts.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>