scripting: drop CallFunction::isqueued flag
The flag is actually not required because we never return from execute() before the function was
either successfully executed or ended up in error state.

CallFunction::valid() should still return true after execute() returned, so it must not be cleared
in executeAndDispose() if ProgramInterface::execute() returned false and the function is either done
or in error.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>