mqueue: read with timeout from message queues
This is another attempt to fix the blocking reads if one end of a message queue
connections has already been disconnected. There is no guarantee that a simple
mq_receive() call will return immediately after select() indicated that new data
is available. The timeout should not hurt in normal situations because the dispatcher
simply calls mqRead() again in the next cycle of the loop in Dispatcher::loop() if
the channel has not been destroyed in the meantime.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>