Introduce new connection introspection node types RemoteNode and SharedConnectionNode
For remote connections it is not always possible to get hands on a pointer to a PortInterface
instance because in some cases the port is a temporary only. Instead, remote ports are
represented by RemoteNode instances now in the connection graph, which only store a qualified
or non-qualified name of the remote port, but no pointer.

Shared connections are now represented with their own node type, SharedConnectionNode, which
can have multiple input and output connections. In this case the node does not represent a
connection endpoint, but the connection object itself.