Use a monotonic clock source whenever possible (for the gnulinux target)
The absolute time returned by rtos_get_time_ns() or RTT::os::TimeService::getNSecs() was generated from
a wall clock (time of day), which is not monotonic and can jump in certain cases, e.g. due to an NTP update.
This time source is used in several internal and public API functions to implement timers or to wait on a
condition variable, mutex or semaphore with a timeout. In all these cases non-steady system time can have
undesirable side effects, which are not expected in a Real-Time Toolkit.

There have been several attempts to fix or at least mitigate that problem:
- http://bugs.orocos.org/show_bug.cgi?id=735
- https://github.com/orocos-toolchain/rtt/pull/129
- https://github.com/orocos-toolchain/rtt/pull/138

This patch introduces some potentially breaking semantic changes, with only minimal changes to the
public API:

- Use clock_gettime(CLOCK_MONOTONIC, ...) to implement rtos_get_time_ns() for the gnulinux target.
  This change has a major impact on all usages of that method, rtos_get_ticks() or the higher-level
  methods RTT::os::TimeService::getNSecs() and RTT::os::TimeService::getTicks(), whose return values
  cannot be interpreted as wall time anymore.
- Consequently remove the rtos_get_time_monotonic_ns() function introduced in
  https://github.com/orocos-toolchain/rtt/pull/138.
- Drop RTT::os::Semaphore::waitUntil() and the underlying OS abstraction functions.
  The only usage within the RTT core was in RTT::os::Timer, which can be rewritten using a
  condition variable. POSIX semaphores do not support waiting with an absolute timeout from the
  monotonic clock.
- Add new OS abstraction functions for waiting on a mutex lock with a relative timeout and use them
  to implement RTT::os::Mutex::timedlock(). POSIX does not provide a pthread_mutexattr_setclock()
  method similar to pthread_condattr_setclock(). For that reason rtos_mutex_lock_until() and
  rtos_mutex_trylock_for() have to fall back to CLOCK_REALTIME and may still be affected by time
  adjustments while waiting.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>