tests: removed special case for high-priority real-time threads in testAffinity
The condition to execute that part was always evaluating to false because it compared
a scheduler type with a priority. The actual test case is invalid because even for a
real-time thread with the highest priority there is no guarantee that it executes at
least once between the start and stop call.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>