tests: fixed some newer unit tests for the xenomai target (Xenomai 2)
Calls to usleep from a SCHED_FIFO thread triggers the context switch signal handler.

CPU affinitiy can only be set when a thread is created with xenomai, but getCPUAffinity()
returns ~0 unconditionally. TimerThread ignores the CPU affinity if getCPUAffinity() is
not supported for the respective target and the test case would fail.

Thread::getPid() and rtos_task_get_pid() is not supported in xenomai either.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>