cmake: Standardize output executable name with orocos_configure_executable
Using orocos_executable() results in an executable that *is* suffixed
with the OROCOS target, but using orocos_configure_executable() results
in an executable that is *not* suffixed with the OROCOS target.

To reduce confusion standardize so that calling either function results
in an executable that *is* suffixed with the OROCOS target.