cmake: move appending to exported libraries and library dirs back to the full macros
The installation directory (AC_INSTALL_DIR) is unknown in the
orocos_configure_xxx() macros because installation is handled by the user.

All the *_EXPORTED_* variables are only needed to find the libraries and include
directories of one package when using orocos_find_package() or orocos_use_package()
in another package that is built within the same build-space, e.g. when using
catkin_make. In all other cases the libraries and their paths are written to and
loaded from the pkg-config (.pc) files, either in the devel-space (for isolated
builds without installation, e.g. catkin_make_isolated or catkin_tools) or
installed.