types: fix type support for carray<T> and add new method TypeInfo::buildPart()
Method TemplateValueFactory<T>::buildReference(void* ptr) is supposed to return a data source that
references the memory addressed by ptr. It is assumed that the pointer points to an instance of T
(or DataType). But for DataType = carray<T> this is not the case if ptr actually points to the first
entry of the C-array (or boost::array or std::array).

To solve this, we need a sizehint in order to return a ValueDataSource<carray<T>> instance instead,
which is initialized with the given pointer and the sizehint. For the overloads without a sizehint
there is no way to return a carray<T> data source with the correct size (which might be even zero!).

This problem also affected method
StructTypeInfo<T>::getMember(internal::Reference* ref, base::DataSourceBase::shared_ptr item, const std::string& name)
which calls type_discovery<T>::referenceMember() internally.

Related to that, to produce data sources for parts of data structures referenced by another data
source, a new method TypeInfo::buildPart() and respective implementations in TemplateValueFactory<T>
and CArrayTypeInfo<T> are needed. PartDataSource<T> was already used for the type discovery in
StructTypeInfo<T>, but there was no way to create a new PartDataSource<T> instance from a TypeInfo
of the member type and a type-punned pointer to the data member of a standard-layout struct.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>